+++ examples/window_debug.rs
@@ -78,17 +78,17 @@ fn main() {
                                 .partial_cmp(&area(b.size()))
                                 .expect("NaN in video mode size")
                         }) {
-                            window.set_fullscreen(Some(Fullscreen::Exclusive(mode)));
+                            window.set_fullscreen(Some(Fullscreen::Exclusive(mode))).unwrap();
                         } else {
                             eprintln!("no video modes available");
                         }
                     }
                     VirtualKeyCode::F => {
                         if window.fullscreen().is_some() {
-                            window.set_fullscreen(None);
+                            window.set_fullscreen(None).unwrap();
                         } else {
                             let monitor = window.current_monitor();
-                            window.set_fullscreen(Some(Fullscreen::Borderless(monitor)));
+                            window.set_fullscreen(Some(Fullscreen::Borderless(monitor))).unwrap();
                         }
                     }
                     #[cfg(waiting_for_set_minimized)]